refactor: changed name for plansApi.ts
